Update /etc/init for anansi, hetzner, pi, shelob, xi and zeta
If filesystem checks fail, continue with a read-only root filesystem so the
machine can be repaired over ssh. In this case, use syslogd -k if we would
normally log to /var/log, and don't attempt to start user init scripts.

Run ntpd directly (without daemon await-net) now the CLOCK_MONOTONIC bug
is fixed so it starts up reliably.

Wait for /dev/log to exist with daemon -w before launching iwd on pi, xi
and zeta. It will fail to log at all if it wins the race against syslogd.